完成学院分析看板

+++ ruoyi-teacher/src/main/java/com/ruoyi/teacher/domain/vo/CollegeRanking.java
@@ -12,5 +12,7 @@ public class CollegeRanking {
     /**学院名称*/
     private String collegeName;
     /**完成率*/
-    private Long CompletionRate;
+    private Double CompletionRate;
+
+
 }
